Improved add/remove/found devices debug messages.
	* src/libtracker-common/tracker-hal.c: Improved add/remove/found
	devices debug messages. 

	* src/trackerd/tracker-status.c: (battery_percentage_cb): Fixed
	message for battery percentage left so we don't always print 0%.

svn path=/trunk/; revision=3116

+++ src/trackerd/tracker-status.c
@@ -409,8 +409,8 @@ battery_percentage_cb (GObject    *object,
 
 	/* FIXME: This could be a configuration option */
 	if (battery_in_use) {
-		g_message ("Battery percentage is now %d%%",
-			   (gint) percentage * 100);
+		g_message ("Battery percentage is now %.0f%%",
+			   percentage * 100);
 		
 		if (percentage <= 0.05) {
 			/* Running on low batteries, stop indexing for now */
